@@ -7,6 +7,7 @@ import (
 	"sync"
 	"testing"
 
+	"github.com/AdguardTeam/dnsproxy/internal/dnsproxytest"
 	"github.com/AdguardTeam/dnsproxy/upstream"
 	"github.com/AdguardTeam/golibs/logutil/slogutil"
 	"github.com/AdguardTeam/golibs/netutil"
@@ -20,8 +21,8 @@ const ipv4OnlyFqdn = "ipv4.only."
 
 func TestDNS64Race(t *testing.T) {
 	ans := newRR(t, ipv4OnlyFqdn, dns.TypeA, 3600, net.ParseIP("1.2.3.4"))
-	ups := &fakeUpstream{
-		onExchange: func(req *dns.Msg) (resp *dns.Msg, err error) {
+	ups := &dnsproxytest.FakeUpstream{
+		OnExchange: func(req *dns.Msg) (resp *dns.Msg, err error) {
 			resp = (&dns.Msg{}).SetReply(req)
 			if req.Question[0].Qtype == dns.TypeA {
 				resp.Answer = []dns.RR{dns.Copy(ans)}
@@ -29,13 +30,13 @@ func TestDNS64Race(t *testing.T) {
 
 			return resp, nil
 		},
-		onAddress: func() (addr string) { return "fake.address" },
-		onClose:   func() (err error) { return nil },
+		OnAddress: func() (addr string) { return "fake.address" },
+		OnClose:   func() (err error) { return nil },
 	}
-	localUps := &fakeUpstream{
-		onExchange: func(_ *dns.Msg) (_ *dns.Msg, _ error) { panic("not implemented") },
-		onAddress:  func() (addr string) { return "fake.address" },
-		onClose:    func() (err error) { return nil },
+	localUps := &dnsproxytest.FakeUpstream{
+		OnExchange: func(_ *dns.Msg) (_ *dns.Msg, _ error) { panic("not implemented") },
+		OnAddress:  func() (addr string) { return "fake.address" },
+		OnClose:    func() (err error) { return nil },
 	}
 
 	dnsProxy := mustNew(t, &Config{
@@ -182,8 +183,8 @@ func TestProxy_Resolve_dns64(t *testing.T) {
 
 	pt := testutil.PanicT{}
 	newUps := func(answers answerMap) (u upstream.Upstream) {
-		return &fakeUpstream{
-			onExchange: func(req *dns.Msg) (resp *dns.Msg, err error) {
+		return &dnsproxytest.FakeUpstream{
+			OnExchange: func(req *dns.Msg) (resp *dns.Msg, err error) {
 				q := req.Question[0]
 				require.Contains(pt, answers, q.Qtype)
 
@@ -196,22 +197,22 @@ func TestProxy_Resolve_dns64(t *testing.T) {
 
 				return resp, nil
 			},
-			onAddress: func() (addr string) { return "fake.address" },
-			onClose:   func() (err error) { return nil },
+			OnAddress: func() (addr string) { return "fake.address" },
+			OnClose:   func() (err error) { return nil },
 		}
 	}
 
 	localRR := newRR(t, ptr64Domain, dns.TypePTR, 3600, pointedDomain)
-	localUps := &fakeUpstream{
-		onExchange: func(req *dns.Msg) (resp *dns.Msg, err error) {
+	localUps := &dnsproxytest.FakeUpstream{
+		OnExchange: func(req *dns.Msg) (resp *dns.Msg, err error) {
 			require.Equal(pt, req.Question[0].Name, ptr64Domain)
 			resp = (&dns.Msg{}).SetReply(req)
 			resp.Answer = []dns.RR{localRR}
 
 			return resp, nil
 		},
-		onAddress: func() (addr string) { return "fake.local.address" },
-		onClose:   func() (err error) { return nil },
+		OnAddress: func() (addr string) { return "fake.local.address" },
+		OnClose:   func() (err error) { return nil },
 	}
 
 	testCases := []struct {
