@@ -295,7 +295,7 @@ func BenchmarkDoTUpstream(b *testing.B) {
 		return len(reqChan) == cap(reqChan)
 	}, time.Second, time.Millisecond)
 
-	b.Run("exchange_p", func(b *testing.B) {
+	b.Run("exchange", func(b *testing.B) {
 		b.ResetTimer()
 		b.ReportAllocs()
 
@@ -305,4 +305,12 @@ func BenchmarkDoTUpstream(b *testing.B) {
 			}
 		})
 	})
+
+	// Most recent results:
+	//
+	//	goos: darwin
+	//	goarch: amd64
+	//	pkg: github.com/AdguardTeam/dnsproxy/upstream
+	//	cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
+	//	BenchmarkDoTUpstream/exchange_p-12         	   90169	     13303 ns/op	    2637 B/op	      45 allocs/op
 }
